feat:新增存储领取 NFT 接口， 查询 NFT 接口，查询已投票接口

+++ backend/proposals/models/proposals.go
@@ -49,7 +49,7 @@ func (GetProposalRequest) TableName() string {
 }
 
 type UpdateProposalsRequest struct {
-	ProposalId  int64          `json:"proposal_id"`
+	ProposalId  int            `json:"proposal_id"`
 	Status      string         `json:"status"`
 	Participate string         `json:"participate"`
 	UpdateMask  pq.StringArray `gorm:"type:text[]" json:"update_mask"`
